fix the problem that portrait cannot save
After uploading an portrait, the data format of portraitBuff does not match the actual format. As a result, the string generated by JSON.stringify contains characters that btoa cannot process, make the portrait cannot be recognized and the "refresh iframe" button doesn't work.

To fix this issue, change the type of portraitBuff to string, and appropriate encoding operations should be applied.